Rename package to match lib and sbt plugins. Rename project to reactive-cli
